feat: abstract header flushed out
Implementation for the abstract header class was modified and these
modifications are reflected in sockets. Implementation tests were
written for abstract header. There is no longer a message class  because
all I was doing was just extracting the header and body every time so I
just cut out the middle man